backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1
websockets==12.0
prometheus-client==0.19.0
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
pyotp==2.9.0